requirements.txt
openai~=1.3.3
pandas~=2.2.2
numpy~=1.26.4
leidenalg~=0.10.2
igraph~=0.11.6
python-dotenv~=1.0.1
tqdm~=4.66.4
tiktoken~=0.7.0
pytesseract~=0.3.10
pillow~=10.4.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
markdown~=3.6
torch~=2.4.0
transformers~=4.43.4
jinja2~=3.1.4
scikit-learn~=1.4.2
